+++ packages/vda-reward-client/CHANGELOG.md
@@ -1,11 +1,13 @@
 
-2024-07-09 (v)
+2024-07-10 (v)
 -------------------
 
 - Add a dependency to the `@verida/vda-client-base`
 - Updated test codes
 - Added `Test.md` file
 - Updated `.env.example` file
+- Added `owner()` function
+- Update test script to let general users run without adding initial data
 
 2024-02-09 (v3.0.0)
 -------------------